requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
spacy>=3.7.2
numpy>=1.24.0
tqdm>=4.66.1
ollama>=0.1.6
pytest>=7.4.0 